+++ src/lib.rs
@@ -2867,7 +2867,7 @@ impl Johnny {
         Ok(v.message_processed())
     }
 
-    pub fn verify_file_and_extract(&self, filepath: Vec<u8>, output: Vec<u8>) -> Result<bool> {
+    pub fn verify_and_extract_file(&self, filepath: Vec<u8>, output: Vec<u8>) -> Result<bool> {
         let p = P::new();
         let vh = VHelper::new(&self.cert);
         let path = Path::new(str::from_utf8(&filepath[..])?);