+++ johnnycanencrypt/__init__.py
@@ -436,12 +436,10 @@ class KeyStore:
         :param other: Either fingerprint as str or `Key` object.
         :returns: boolean result
         """
-        fingerprint = ""
-        if type(other) == str:
-            assert isinstance(other, str)
+        fingerprint: str = ""
+        if isinstance(other, str):
             fingerprint = other
-        elif type(other) == Key:
-            assert isinstance(other, Key)
+        elif isinstance(other, Key):
             fingerprint = other.fingerprint
         try:
             if self.get_key(fingerprint):
@@ -986,14 +984,12 @@ class KeyStore:
 
         :param key: Either str representation of the fingerprint or a Key object
         """
-        if type(key) == str:
-            assert isinstance(key, str)
+        if isinstance(key, str):
             fingerprint = key
-        elif type(key) == Key:
-            assert isinstance(key, Key)
+        elif isinstance(key, Key):
             fingerprint = key.fingerprint
         else:
-            raise TypeError(f"Wrong datatype for {key}")
+            raise TypeError(f"Wrong datatype for {str(key)}")
 
         if not fingerprint in self:
             raise KeyNotFoundError(
@@ -1004,18 +1000,18 @@ class KeyStore:
             cursor = con.cursor()
             cursor.execute("DELETE FROM keys where fingerprint=?", (fingerprint,))
 
-    def _find_keys(self, keys):
+    def _find_keys(self, keys: List[Union[str, Key]]):
         "To find all the key paths"
         final_keys = []
         for k in keys:
-            if type(k) == str:  # Means fingerprint
+            if isinstance(k, str):  # Means fingerprint
                 key = self.get_key(k)
                 final_keys.append(key.keyvalue)
             else:
                 final_keys.append(k.keyvalue)
         return final_keys
 
-    def encrypt(self, keys, data, outputfile="", armor=True):
+    def encrypt(self, keys: Union[List[Union[str, Key]], Union[str, Key]], data: Union[str, bytes], outputfile: Union[str, bytes]="", armor=True):
         """Encrypts the given data with the list of keys and returns the output.
 
         :param keys: List of fingerprints or Key objects
@@ -1023,7 +1019,7 @@ class KeyStore:
         :param outputfile: If provided the output will be wriiten in the location.
         :param armor: Default is True, for armored output.
         """
-        if type(keys) != list:
+        if not isinstance(keys, list):
             finalkeys = [
                 keys,
             ]
@@ -1031,7 +1027,7 @@ class KeyStore:
             finalkeys = keys
         final_key_paths = self._find_keys(finalkeys)
         # Check if we return data
-        if type(data) == str:
+        if isinstance(data, str):
             finaldata = data.encode("utf-8")
         else:
             finaldata = data
@@ -1039,7 +1035,7 @@ class KeyStore:
             return encrypt_bytes_to_bytes(final_key_paths, finaldata, armor)
 
         # For encryption to a file
-        if type(outputfile) == str:
+        if isinstance(outputfile, str):
             encrypted_file = outputfile.encode("utf-8")
         else:
             encrypted_file = outputfile
@@ -1047,14 +1043,14 @@ class KeyStore:
         encrypt_bytes_to_file(final_key_paths, finaldata, encrypted_file, armor)
         return True
 
-    def decrypt(self, key, data, password=""):
+    def decrypt(self, key: Union[str, Key], data, password=""):
         """Decrypts the given bytes and returns plain text bytes.
 
         :param key: Fingerprint or secret Key object
         :param data: Encrypted data in bytes.
         :param password: Password for the secret key
         """
-        if type(key) == str:  # Means we have a fingerprint
+        if isinstance(key, str):  # Means we have a fingerprint
             k = self.get_key(key)
         else:
             k = key
@@ -1095,7 +1091,7 @@ class KeyStore:
             if not os.path.exists(inputfilepath):
                 raise FileNotFoundError(f"{inputfilepath} can not be found.")
 
-        if type(keys) != list:
+        if not isinstance(keys, list):
             finalkeys = [
                 keys,
             ]
@@ -1104,7 +1100,7 @@ class KeyStore:
         final_key_paths = self._find_keys(finalkeys)
 
         # For encryption to a file
-        if type(outputfilepath) == str:
+        if isinstance(outputfilepath, str):
             encrypted_file = outputfilepath.encode("utf-8")
         else:
             encrypted_file = outputfilepath
@@ -1115,7 +1111,7 @@ class KeyStore:
             encrypt_filehandler_to_file(final_key_paths, fh, encrypted_file, armor)
         return True
 
-    def decrypt_file(self, key, encrypted_path, outputfile, password=""):
+    def decrypt_file(self, key: Union[str, Key], encrypted_path, outputfile, password=""):
         """Decryptes the given file to the output path.
 
         :param key: Fingerprint or secret Key object
@@ -1124,12 +1120,12 @@ class KeyStore:
         :param password: Password for the secret key
         """
         use_filehandler = False
-        if type(key) == str:  # Means we have a fingerprint
+        if isinstance(key, str):  # Means we have a fingerprint
             k = self.get_key(key)
         else:
             k = key
 
-        if type(encrypted_path) == str:
+        if isinstance(encrypted_path, str):
             inputfile = encrypted_path.encode("utf-8")
         elif isinstance(encrypted_path, bytes):
             inputfile = encrypted_path
@@ -1137,7 +1133,7 @@ class KeyStore:
             fh = encrypted_path
             use_filehandler = True
 
-        if type(outputfile) == str:
+        if isinstance(outputfile, str):
             outputpath = outputfile.encode("utf-8")
         else:
             outputpath = outputfile
@@ -1159,7 +1155,7 @@ class KeyStore:
         else:
             return jp.decrypt_filehandler(fh, outputpath, password)
 
-    def sign_detached(self, key, data, password):
+    def sign_detached(self, key: Union[str, Key], data: Union[str, bytes], password):
         """Signs the given data with the key.
 
         :param key: Fingerprint or secret Key object
@@ -1168,12 +1164,12 @@ class KeyStore:
 
         :returns: The signature as string
         """
-        if type(key) == str:  # Means we have a fingerprint
+        if isinstance(key, str):  # Means we have a fingerprint
             k = self.get_key(key)
         else:
             k = key
 
-        if type(data) == str:
+        if isinstance(data, str):
             data = data.encode("utf-8")
 
         if k.keytype == KeyType.PUBLIC and k.oncard is not None:
@@ -1184,7 +1180,7 @@ class KeyStore:
         jp = Johnny(k.keyvalue)
         return jp.sign_bytes_detached(data, password)
 
-    def verify(self, key, data, signature: Optional[str]) -> bool:
+    def verify(self, key: Union[str, Key], data: Union[str, bytes], signature: Optional[str]) -> bool:
         """Verifies the given data and the signature
 
         :param key: Fingerprint or public Key object
@@ -1193,12 +1189,12 @@ class KeyStore:
 
         :returns: Boolean
         """
-        if type(key) == str:  # Means we have a fingerprint
+        if isinstance(key, str):  # Means we have a fingerprint
             k = self.get_key(key)
         else:
             k = key
 
-        if type(data) == str:
+        if isinstance(data, str):
             data = data.encode("utf-8")
         jp = Johnny(k.keyvalue)
 
@@ -1207,7 +1203,7 @@ class KeyStore:
         else:
             return jp.verify_bytes(data)
 
-    def sign_file(self, key, filepath, outputpath, password, cleartext=False) -> bool:
+    def sign_file(self, key: Union[str, Key], filepath: Union[str, bytes], outputpath: Union[str, bytes], password, cleartext=False) -> bool:
         """Signs the given input file with key and saves in the outputpath.
 
         :param key: Fingerprint or secret Key object, public key in case card based operation.
@@ -1219,17 +1215,17 @@ class KeyStore:
         :returns: Boolean result of the signing operation.
         """
         signature = ""
-        if type(key) == str:  # Means we have a fingerprint
+        if isinstance(key, str):  # Means we have a fingerprint
             k = self.get_key(key)
         else:
             k = key
 
-        if type(filepath) == str:
+        if isinstance(filepath, str):
             filepath_in_bytes = filepath.encode("utf-8")
         else:
             filepath_in_bytes = filepath
 
-        if type(outputpath) == str:
+        if isinstance(outputpath, str):
             outputpath_in_bytes = outputpath.encode("utf-8")
         else:
             outputpath_in_bytes = outputpath
@@ -1251,23 +1247,23 @@ class KeyStore:
 
         return result
 
-    def sign_file_detached(self, key, filepath, password, write=False):
+    def sign_file_detached(self, key: Union[str, Key], filepath: Union[str, bytes], password: str, write=False):
         """Signs the given data with the key. It also writes filename.asc in the same directory of the file as the signature if write value is True.
 
         :param key: Fingerprint or secret Key object
         :param filepath: str value of the path to the file.
-        :param password: Password of the secret key file.
+        :param password: Password of the secret key file as str.
         :param write: boolean value (default False), determines if we should write the signature to a file.
 
         :returns: The signature as string
         """
         signature = ""
-        if type(key) == str:  # Means we have a fingerprint
+        if isinstance(key, str):  # Means we have a fingerprint
             k = self.get_key(key)
         else:
             k = key
 
-        if type(filepath) == str:
+        if isinstance(filepath, str):
             filepath_in_bytes = filepath.encode("utf-8")
         else:
             filepath_in_bytes = filepath
@@ -1283,13 +1279,13 @@ class KeyStore:
 
         # Now check if we have to write the file on disk
         if write:
-            sig_file_name = filepath + ".asc"
+            sig_file_name = f'{filepath_in_bytes.decode("utf-8")}.asc'
             with open(sig_file_name, "w") as fobj:
                 fobj.write(signature)
 
         return signature
 
-    def verify_file_detached(self, key, filepath, signature_path):
+    def verify_file_detached(self, key: Union[str, Key], filepath: Union[str, bytes], signature_path):
         """Verifies the given filepath based on the signature file.
 
         :param key: Fingerprint or public Key object
@@ -1298,7 +1294,7 @@ class KeyStore:
 
         :returns: Boolean
         """
-        if type(key) == str:  # Means we have a fingerprint
+        if isinstance(key, str):  # Means we have a fingerprint
             k = self.get_key(key)
         else:
             k = key
@@ -1308,18 +1304,18 @@ class KeyStore:
                 f"The signature file at {signature_path} is missing."
             )
         if not os.path.exists(filepath):
-            raise FileNotFoundError(f"The file at {filepath} is missing.")
+            raise FileNotFoundError(f"The file at {str(filepath)} is missing.")
 
         # Let us read the signature
         with open(signature_path, "rb") as fobj:
             signature_in_bytes = fobj.read()
 
-        if type(filepath) == str:
+        if isinstance(filepath, str):
             filepath = filepath.encode("utf-8")
         jp = Johnny(k.keyvalue)
         return jp.verify_file_detached(filepath, signature_in_bytes)
 
-    def verify_file(self, key, filepath):
+    def verify_file(self, key: Union[str, Key], filepath):
         """Verifies the given filepath.
 
         :param key: Fingerprint or public Key object
@@ -1327,18 +1323,71 @@ class KeyStore:
 
         :returns: Boolean
         """
-        if type(key) == str:  # Means we have a fingerprint
+        if isinstance(key, str):  # Means we have a fingerprint
             k = self.get_key(key)
         else:
             k = key
 
         if not os.path.exists(filepath):
-            raise FileNotFoundError(f"The file at {filepath} is missing.")
+            raise FileNotFoundError(f"The file at {str(filepath)} is missing.")
 
-        if type(filepath) == str:
-            filepath = filepath.encode("utf-8")
+        if isinstance(filepath, str):
+            input_filepath = filepath.encode("utf-8")
+        else:
+            input_filepath = filepath
+
         jp = Johnny(k.keyvalue)
-        return jp.verify_file(filepath)
+        return jp.verify_file(input_filepath)
+
+    def verify_and_extract_bytes(self, key: Union[str, Key], data: Union[str, bytes]) -> bytes:
+        """Verifies the given data and returns the acutal data.
+
+        :param key: Fingerprint or public Key object.
+        :param data: Data to be signed.
+
+        :returns: bytes
+        """
+        if isinstance(key, str):  # Means we have a fingerprint
+            k = self.get_key(key)
+        else:
+            k = key
+
+        if isinstance(data, str):
+            data = data.encode("utf-8")
+        jp = Johnny(k.keyvalue)
+
+        return jp.verify_and_extract_bytes(data)
+
+    def verify_and_extract_file(self, key: Union[str, Key], filepath: Union[str, bytes], output: bytes) -> bool:
+        """Verifies the given signed file and saves the actual data in output.
+
+        :param key: Fingerprint or public Key object.
+        :param filepath: Signed file as bytes.
+        :param output: Output path for the original content.
+
+        :returns: bool
+        """
+        if isinstance(key, str):  # Means we have a fingerprint
+            k = self.get_key(key)
+        else:
+            k = key
+
+        if not os.path.exists(filepath):
+            raise FileNotFoundError(f"The file at {str(filepath)} is missing.")
+
+        if isinstance(filepath, str):
+            input_filepath = filepath.encode("utf-8")
+        else:
+            input_filepath = filepath
+
+        if isinstance(output, str):
+            outputpath = output.encode("utf-8")
+        else:
+            outputpath = output
+        jp = Johnny(k.keyvalue)
+
+        return jp.verify_and_extract_file(input_filepath, outputpath)
+
 
     def fetch_key_by_fingerprint(self, fingerprint: str):
         """Fetches key from keys.openpgp.org based on the fingerprint.